# swagger_ui/openapi_ms1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Investment Data Service API",
    "description": "Documentação completa das rotas do microserviço investment_data_service",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Servidor local"
    }
  ],
  "paths": {
...
    "/admin/transacoes/{id_cliente}": {
      "get": {
        "summary": "Listar transações de um cliente",
        "tags": ["Admin"],
        "parameters": [
          {
            "name": "id_cliente",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Transações encontradas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id_transacao": { "type": "integer" },
                      "cliente_id": { "type": "string" },
                      "tipo": { "type": "string" },
                      "valor": { "type": "number" },
                      "data": { "type": "string" },
                      "descricao": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/investidor/{id_cliente}": {
      "get": {
        "summary": "Buscar investidor pelo ID do cliente",
        "tags": ["Investidor"],
        "parameters": [
          {
            "name": "id_cliente",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Investidor encontrado"
          },
          "404": {
            "description": "Investidor não encontrado"
          }
        }
      }
    },
...
  },  
...
}